- Some inlining optimizations.
git-svn-id: http://svn.code.sf.net/p/qtractor/code_attic/trunk@1708 304e7ffa-9037-4501-99c0-529757c9bbcc

+++ src/qtractorAudioSndFile.cpp
@@ -142,7 +142,7 @@ bool qtractorAudioSndFile::seek ( unsigned long iOffset )
 #ifdef DEBUG_0
 	qDebug("qtractorAudioSndFile::seek(%d)", iOffset);
 #endif
-	return (::sf_seek(m_pSndFile, iOffset, SEEK_SET) == iOffset);
+	return (::sf_seek(m_pSndFile, iOffset, SEEK_SET) == long(iOffset));
 }
